requirements.txt
pygame-ce>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

